# backend/internal/handlers/openapi.yaml
openapi: 3.0.3
info:
  title: RallyUp API
  description: RallyUp Badminton Club management API — member registration, session scheduling, RSVP tracking, and notifications.
  version: 1.0.0
servers:
  - url: /api
    description: API root
...
security:
  - bearerAuth: []
paths:
  /club:
    get:
      tags: [Public]
      summary: Get club information
      operationId: getClub
      security: []
      responses:
        "200":
          description: Club info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /users:
    get:
      tags: [Users]
      summary: List approved club members
      description: Requires approved membership.
      operationId: listMembers
      responses:
        "200":
          description: List of approved members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /admin/join-requests:
    get:
      tags: ["Admin: Members"]
      summary: List pending join requests
      operationId: listJoinRequests
      responses:
        "200":
          description: List of pending users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/join-requests/{id}/approve:
    post:
      tags: ["Admin: Members"]
      summary: Approve a join request
      operationId: approveJoinRequest
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Approved user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/join-requests/{id}/reject:
    post:
      tags: ["Admin: Members"]
      summary: Reject a join request
      operationId: rejectJoinRequest
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Rejected user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users:
    get:
      tags: ["Admin: Members"]
      summary: List every member, whatever their status
      description: >
        The admin's view of the roll. Unlike `GET /users`, which is the club's
        list of approved members, this includes pending, rejected and removed
        rows, because those are what an admin acts on.
      operationId: adminListMembers
      responses:
        "200":
          description: Every user row
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: ["Admin: Members"]
      summary: Invite a member who has not signed up yet
      description: >
        Creates an approved member ahead of their first sign-in. Sign-in is
        Google OAuth, so there is no account to create on their behalf; what this
        row does is reserve the identity, carrying the membership, the role and a
        ledger account. The member is chargeable immediately — they can be RSVP'd
        into sessions and settled against. When they first sign in with a
        verified email matching this one, `POST /auth/callback` adopts the row
        rather than creating a second one, so they arrive already approved.


        No email is sent; tell them to sign in however you normally reach them.
      operationId: inviteMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: The invited member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}:
    put:
      tags: ["Admin: Members"]
      summary: Update a member's details
      description: >
        Omitted fields are left alone, so the form can send only what changed.


        `email` is editable only while the row is still an unclaimed invite. Once
        a real Auth0 subject is bound to it, the address belongs to the identity
        provider: changing it here would not change how they sign in, and would
        silently move them in or out of the ADMIN_EMAIL auto-promotion rule.
      operationId: updateMember
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: The updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No member with that id
    delete:
      tags: ["Admin: Members"]
      summary: Remove a member from the club
      description: >
        Revokes access by setting `membership_status` to `removed`. This is not a
        delete — the row and its ledger history stay, and `POST
        /admin/users/{id}/reinstate` undoes it.


        Refused with 400 when the member still has a non-zero balance (settle up
        with a top-up or withdrawal first), when they are the club's only admin,
        or when they are the caller. Their RSVPs for upcoming sessions are
        cancelled, promoting anyone waiting behind them.
      operationId: removeMember
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: The removed member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No member with that id
  /admin/users/{id}/reinstate:
    post:
      tags: ["Admin: Members"]
      summary: Return a removed member to the club
      operationId: reinstateMember
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: The reinstated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No member with that id
  /admin/users/{id}/role:
    put:
      tags: ["Admin: Members"]
      summary: Update a user's role
      operationId: updateUserRole
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/sessions:
    post:
      tags: ["Admin: Sessions"]
      summary: Create a session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/sessions/{id}:
    put:
      tags: ["Admin: Sessions"]
      summary: Update a session
      operationId: updateSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionRequest"
      responses:
        "200":
          description: Session updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: ["Admin: Sessions"]
      summary: Delete a session
      operationId: deleteSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          $ref: "#/components/responses/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/sessions/{id}/cancel:
    post:
      tags: ["Admin: Sessions"]
      summary: Cancel a session
      operationId: cancelSession
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelSessionRequest"
      responses:
        "200":
          description: Cancelled session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /admin/club:
    put:
      tags: ["Admin: Club"]
      summary: Update club settings
      operationId: updateClub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClubRequest"
      responses:
        "200":
          description: Updated club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 JWT token
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Message:
      description: Success message
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
    Club:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        venue_name:
          type: string
        venue_address:
          type: string
        notifications_paused:
          type: boolean
          description: Discards notifications while setup or data review is in progress.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        auth0_id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        nickname:
          type: string
          description: >
            What the club calls this member, chosen by them or corrected by an
            admin. Displayed in place of `name` wherever it is set — member
            lists, RSVPs, balances and settlement breakdowns. When it is empty
            the display name falls back to the member's first name, so this
            being blank does not mean they show up under their full legal name.
        profile_picture:
          type: string
        phone_number:
          type: string
        role:
          type: string
          enum: [pending, player, admin]
        is_player:
          type: boolean
        membership_status:
          type: string
          enum: [pending, approved, rejected, removed]
          description: >
            `removed` is a member an admin has taken out of the club. The row is
            never deleted: RSVPs, ledger entries and settlements reference it,
            and the ledger is append-only.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        session_date:
          type: string
          format: date
        start_time:
          type: string
          description: "HH:MM format"
          example: "18:00"
        end_time:
          type: string
          description: "HH:MM format"
          example: "20:00"
        courts:
          type: integer
          minimum: 1
          maximum: 3
        max_players:
          type: integer
        rsvp_deadline:
          type: string
          format: date-time
        is_recurring:
          type: boolean
        recurring_day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          nullable: true
          description: "0=Sunday, 1=Monday, ..., 6=Saturday"
        recurring_parent_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [open, closed, cancelled]
        cancellation_reason:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        rsvps:
          type: array
          items:
            $ref: "#/components/schemas/RSVP"
        creator:
          $ref: "#/components/schemas/User"
...
    UpdateRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [pending, player, admin]
    InviteMemberRequest:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          format: email
          description: >
            Matched case-insensitively against the verified email at first
            sign-in, which is how the invite is claimed.
        name:
          type: string
        nickname:
          type: string
        phone_number:
          type: string
        role:
          type: string
          enum: [player, admin]
          default: player
    UpdateMemberRequest:
      type: object
      description: Every field is optional; omitted fields are left unchanged.
      properties:
        name:
          type: string
        nickname:
          type: string
        phone_number:
          type: string
        email:
          type: string
          format: email
          description: Accepted only while the member has never signed in.
        role:
          type: string
          enum: [pending, player, admin]
        is_player:
          type: boolean
    CreateSessionRequest:
      type: object
      required: [title, session_date, start_time, end_time, courts]
      properties:
        title:
          type: string
        description:
          type: string
        session_date:
          type: string
          format: date
          description: "YYYY-MM-DD"
        start_time:
          type: string
          description: "HH:MM"
          example: "18:00"
        end_time:
          type: string
          description: "HH:MM"
          example: "20:00"
        courts:
          type: integer
          minimum: 1
          maximum: 3
        is_recurring:
          type: boolean
        recurring_day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          nullable: true
        occurrences:
          type: integer
          nullable: true
          description: Number of recurring sessions to create
    UpdateSessionRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        session_date:
          type: string
          format: date
          description: "YYYY-MM-DD"
        start_time:
          type: string
          description: "HH:MM"
        end_time:
          type: string
          description: "HH:MM"
        courts:
          type: integer
          minimum: 1
          maximum: 3
        status:
          type: string
          enum: [open, closed, cancelled]
    CancelSessionRequest:
      type: object
      properties:
        reason:
          type: string
    UpdateClubRequest:
      type: object
      properties:
        notifications_paused:
          type: boolean
          description: Pauses all notification delivery. Does not queue messages for later sending.
        name:
          type: string
        venue_name:
          type: string
        venue_address:
          type: string
...